Add files via upload
Here I try to exploit the fact that there is no temporal correlation between clicks so that I can use different permutations of the same register. For each permutation I have a predicted value of detunning so that I end up with a sampling distribution and can gain statistical insight over the probability distribution.